tests/submission/PhilipTaylor/canvas/size.attributes.parse.zero.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: size.attributes.parse.zero</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.<a href="index.size.attributes.parse.html">parse</a>.zero</h1>
<p class="desc">Parsing of non-negative integers</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.nonnegativeinteger">size.nonnegativeinteger</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="0" id="c" width="0"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Parsing of non-negative integers");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 0, "canvas.width", "0");
_assertSame(canvas.height, 0, "canvas.height", "0");
_assertEqual(window.getComputedStyle(canvas, null).getPropertyValue("width"), "0px", "window.getComputedStyle(canvas, null).getPropertyValue(\"width\")", "\"0px\"");


});
</script>

<div id="log"></div>
</body></html>
